Threads/src/test/java/ru/job4j/concurrent/SingleLockListTest.java
package ru.job4j.concurrent;

import org.junit.Test;

import java.util.Set;
import java.util.TreeSet;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

public class SingleLockListTest {
    @Test
    public void add() throws InterruptedException {
        SingleLockList<Integer> list = new SingleLockList<>();
        Thread first = new Thread(() -> list.add(1));
        Thread second = new Thread(() -> list.add(2));
        first.start();
        second.start();
        first.join();
        second.join();
        Set<Integer> rsl = new TreeSet<>();
        list.iterator().forEachRemaining(rsl::add);
        Set<Integer> expect = new TreeSet<>();
        expect.add(1);
        expect.add(2);
        assertThat(rsl, is(expect));
    }
}
